dev: accounts task printing all named accounts with avax, dai, and defo balances
