+++ .gitignore
@@ -9,6 +9,7 @@
 **/build
 **/cache
 **/coverage
+**/deployments
 **/.coverage_artifacts
 **/.coverage_cache
 **/.coverage_contracts
