+++ package.json
@@ -25,6 +25,7 @@
     "@types/debug": "^4.1.7",
     "@types/mocha": "^9.1.1",
     "@types/node": "^18.0.1",
+    "@types/qrcode-terminal": "^0.12.0",
     "@typescript-eslint/eslint-plugin": "^5.30.5",
     "@typescript-eslint/parser": "^5.30.5",
     "chai": "^4.3.6",
@@ -34,7 +35,6 @@
     "cross-env": "^7.0.3",
     "cz-conventional-changelog": "3.3.0",
     "debug": "^4.3.4",
-    "deploy-eip-1820": "^1.0.0",
     "eslint": "^8.19.0",
     "eslint-config-prettier": "^8.5.0",
     "ethereum-waffle": "^3.4.4",
@@ -43,6 +43,7 @@
     "hardhat-abi-exporter": "^2.10.0",
     "hardhat-deploy": "^0.11.11",
     "hardhat-deploy-ethers": "^0.3.0-beta.13",
+    "hardhat-erc1820": "^0.1.0",
     "hardhat-ethernal": "^1.0.0",
     "hardhat-gas-reporter": "^1.0.8",
     "husky": "^8.0.1",
@@ -50,6 +51,7 @@
     "mocha": "^10.0.0",
     "prettier": "^2.7.1",
     "prettier-plugin-solidity": "^1.0.0-beta.19",
+    "qrcode-terminal": "^0.12.0",
     "shelljs": "^0.8.5",
     "shx": "^0.3.4",
     "solhint": "^3.3.7",
@@ -77,7 +79,7 @@
     "access": "public"
   },
   "scripts": {
-    "clean": "shx rm -rf ./artifacts ./cache ./coverage ./src/types ./coverage.json && yarn typechain",
+    "clean": "shx rm -rf ./artifacts ./deployments ./cache ./coverage ./src/types ./coverage.json && yarn typechain",
     "docs": "solidity-docgen --solc-module solc-0.8 -i contracts -o docs -t templates",
     "commit": "git-cz",
     "compile": "cross-env TS_NODE_TRANSPILE_ONLY=true hardhat compile",
@@ -100,9 +102,11 @@
     "deploy": "hardhat deploy --write true --network ",
     "accounts": "hardhat --network localhost accounts",
     "get-some-dai": "hardhat --network localhost fork:get-some-dai --amount 10",
-    "mint-some-gems": "hardhat --network localhost fork:mint-some-gems --amount 10  ",
-    "jump-in-time-one-year": "hardhat --network localhost fork:jump-in-time --time 1y",
-    "run-fork": "yarn get-some-dai && yarn mint-some-gems && yarn jump-in-time-one-year",
+    "mint-gems": "hardhat --network localhost fork:mint-some-gems --amount",
+    "mint-some-gems": "yarn mint-gems 10",
+    "jump-in-time": "hardhat --network localhost fork:jump-in-time --time",
+    "jump-in-time-1y": "jump-in-time 1y",
+    "run-fork": "yarn get-some-dai && yarn mint-some-gems && yarn jump-in-time-1y",
     "run-fork:fuji": "FORK_TESTNET=true yarn run-fork",
     "test": "HIDE_DEPLOY_LOG=true OPTIMIZER_DISABLED=true hardhat test",
     "gas": "REPORT_GAS=true HIDE_DEPLOY_LOG=true hardhat test",
