+++ tasks/accounts.ts
@@ -1,42 +1,25 @@
-// import { formatUnits } from "@ethersproject/units";
-// import { task } from "hardhat/config";
-// import { HttpNetworkUserConfig } from "hardhat/types";
-// import { THardhatRuntimeEnvironmentExtended } from "helpers/types/THardhatRuntimeEnvironmentExtended";
-// import * as qrcode from "qrcode-terminal";
-// import { findFirstAddress, getAccountData } from "tasks/functions/accounts";
-// import { DEBUG } from "tasks/functions/debug";
-//
-// import { config } from "../hardhat.config";
-// import { getMnemonic } from "./functions/mnemonic";
-//
-// task("account", "Get balance informations for the deployment account.", async (_, hre) => {
-//   const { address } = await getAccountData(getMnemonic());
-//
-//   qrcode.generate(address);
-//   console.log(`‍📬 Deployer Account is ${address}`);
-//   for (const n in config.networks) {
-//     // console.log(config.networks[n],n)
-//     try {
-//       const { url } = config.networks[n] as HttpNetworkUserConfig;
-//       const provider = new hre.ethers.providers.JsonRpcProvider(url);
-//       const balance = await provider.getBalance(address);
-//       console.log(` -- ${n} --  -- -- 📡 `);
-//       console.log(`   balance: ${hre.ethers.utils.formatEther(balance)}`);
-//       console.log(`   nonce: ${await provider.getTransactionCount(address)}`);
-//     } catch (e) {
-//       if (DEBUG) console.log(e);
-//     }
-//   }
-// });
-//
-// task("balance", "Prints an account's balance")
-//   .addPositionalParam("account", "The account's address")
-//   .setAction(async (taskArgs: { account: string }, hre: THardhatRuntimeEnvironmentExtended) => {
-//     const balance = await hre.ethers.provider.getBalance(await findFirstAddress(hre, taskArgs.account));
-//     console.log(formatUnits(balance, "ether"), "ETH");
-//   });
-//
-// task("accounts", "Prints the list of accounts", async (_, { ethers }) => {
-//   const accounts = await ethers.provider.listAccounts();
-//   accounts.forEach((account: any) => console.log(account));
-// });
+import { task } from "hardhat/config";
+
+import ERC20_ABI from "../abi/defo-abi.json";
+import { info } from "../utils/helpers";
+
+task("accounts", "Get the addresses and balance information (AVAX, DEFO, DAI) for the accounts.", async (_, hre) => {
+  const { getNamedAccounts } = hre;
+  const namedAccounts = await getNamedAccounts();
+  const { dai, defoToken } = namedAccounts;
+  info("\n 📡 Querying balances...");
+  const daiContract = await hre.ethers.getContractAt(ERC20_ABI, dai);
+  const defoContract = await hre.ethers.getContractAt(ERC20_ABI, defoToken);
+  const table = await Promise.all(
+    Object.entries(namedAccounts).map(async ([accountName, accountAddress]) => {
+      return {
+        name: accountName,
+        address: accountAddress,
+        AVAX: Number(hre.ethers.utils.formatEther(await hre.ethers.provider.getBalance(accountAddress))),
+        DAI: Number(hre.ethers.utils.formatEther(await daiContract.balanceOf(accountAddress))),
+        DEFO: Number(hre.ethers.utils.formatEther(await defoContract.balanceOf(accountAddress))),
+      };
+    }),
+  );
+  console.table(table);
+});
